requirements.txt
flask
torch
ultralytics
transformers
Pillow
docker
gradio